+++ src/utils/schema/resolvers/Query/all.ts
@@ -1,5 +1,5 @@
 import { GraphQLError } from 'graphql';
-import { Form, Record, ReferenceData, User } from '@models';
+import { Record, ReferenceData, User } from '@models';
 import extendAbilityForRecords from '@security/extendAbilityForRecords';
 import { decodeCursor, encodeCursor } from '@schema/types';
 import getReversedFields from '../../introspection/getReversedFields';
@@ -19,9 +19,14 @@ import checkPageSize from '@utils/schema/errors/checkPageSize.util';
 import { flatten, get, isArray, set } from 'lodash';
 import { accessibleBy } from '@casl/mongoose';
 import { graphQLAuthCheck } from '@schema/shared';
+import NodeCache from 'node-cache';
+import { AppAbility } from '@security/defineUserAbility';
 
 /** Default number for items to get */
 const DEFAULT_FIRST = 25;
+
+/** Ability Cache, based on user id, time to live: 5min */
+const abilityCache = new NodeCache({ stdTTL: 60 * 5, checkperiod: 60 });
 
 // todo: improve by only keeping used fields in the $project stage
 /**
@@ -233,6 +238,7 @@ export default (entityName: string, fieldsByName: any, idsByName: any) =>
     checkPageSize(first);
     try {
       const user: User = context.user;
+      const userId = user._id.toString();
       // Id of the form / resource
       const id = idsByName[entityName];
       // List of form / resource fields
@@ -407,16 +413,25 @@ export default (entityName: string, fieldsByName: any, idsByName: any) =>
       };
 
       // Additional filter from the user permissions
-      const form = await Form.findOne({
-        $or: [{ _id: id }, { resource: id, core: true }],
-      })
-        .select('_id permissions fields')
-        .populate({ path: 'resource', model: 'Resource' });
-      const ability = await extendAbilityForRecords(user, form);
-      set(context, 'user.ability', ability);
-      const permissionFilters = Record.find(
-        accessibleBy(ability, 'read').Record
-      ).getFilter();
+      let permissionFilters;
+      // Try to get ability from cache
+      let ability = abilityCache.get<AppAbility>(userId);
+      if (!ability) {
+        // If not available, build ability
+        ability = await extendAbilityForRecords(user);
+        set(context, 'user.ability', ability);
+        permissionFilters = Record.find(
+          accessibleBy(ability, 'read').Record
+        ).getFilter();
+        // And cache it
+        abilityCache.set(userId, ability);
+      } else {
+        // Update user ability
+        set(context, 'user.ability', ability);
+        permissionFilters = Record.find(
+          accessibleBy(ability, 'read').Record
+        ).getFilter();
+      }
 
       // Finally putting all filters together
       const filters = {